Create README.md
A python oriented home automation system that regulates the temperature of a room or office automatically based on the presence or non presence of a user in an office or room during the different times of a day. The motion sensor will detect the presence of a user, and after sometime (x minutes) of non presence, the heating system will be off until some minimum temperature reaches, during the working hours, while it totally turns off the temperature during the out of work time in order to save the energy consumption. Demonstrated using a python script that simulates the sensors (publishers) and clients (subscribers) and the things Speak (https://api.thingspeak.com/update?api_key=N9QWT0PB2WI6IBTO).